refactor: split shared.rs, observability.rs, and responses.rs into submodules (#874)
Split three large files (>600 lines each) into focused submodules following
the Single Responsibility Principle:

- src/analysis/shared.rs → shared/{mod,timing,metadata,gpu_info}.rs
- src/observability.rs → observability/{mod,phase_timer,gpu_metrics,profile}.rs
- src/ffi_types/responses.rs → responses/{mod,analysis,gpu,export}.rs

All public types remain accessible via the same import paths through
mod.rs re-exports. No new files exceed 300 lines.
