+++ src/analysis/mod.rs
@@ -11,7 +11,7 @@
 //! - `scoring/` - Scoring, confidence, and validation (weights, confidence, etc.)
 //!
 //! ## Core modules (root level)
-//! - `shared.rs` - Common types, result structures, diagnostics
+//! - `shared/` - Common types, result structures, diagnostics (timing, metadata, gpu_info)
 //! - `synapse/` - Synapse analysis pipeline
 //! - `neuron/` - Neuron analysis (evaluation, post-processing, preparation)
 //! - `gpu/` - GPU infrastructure (GpuAnalyzer, GpuWorkQueue)
