+++ src/ffi_types/responses/mod.rs
@@ -1,0 +1,208 @@
+//! FFI response structs (output to NEAT-AI).
+//!
+//! All `*Output` types and supporting diagnostic/metadata types used at the
+//! FFI boundary for serialising JSON responses.
+//!
+//! Organised into focused submodules:
+//! - [`analysis`] — Analysis output, metadata, and diagnostic types
+//! - [`gpu`] — GPU check output and timing/adapter JSON types
+//! - [`export`] — Export, merge, read, and calibration output types
+
+pub mod analysis;
+pub mod export;
+pub mod gpu;
+
+// Re-export all public types for backward compatibility.
+pub use analysis::*;
+pub use export::*;
+pub use gpu::*;
+
+use serde::Serialize;
+
+use super::DiscoveryErrorKind;
+
+use super::{CoordinatedStructuralCandidateJson, RankedNeuronJson, RemovalCandidateJson};
+
+use crate::analysis as analysis_mod;
+
+/// JSON output from record_discovery function
+#[derive(Debug, Serialize)]
+pub struct RecordDiscoveryOutput {
+    pub success: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub temp_dir: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub file: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+    /// Structured error classification for retry decisions (Issue #651).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error_kind: Option<DiscoveryErrorKind>,
+    /// Whether this error is typically worth retrying (Issue #651).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub retryable: Option<bool>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GetVersionOutput {
+    pub success: bool,
+    pub version: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+    /// Structured error classification for retry decisions (Issue #651).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error_kind: Option<DiscoveryErrorKind>,
+    /// Whether this error is typically worth retrying (Issue #651).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub retryable: Option<bool>,
+}
+
+#[derive(Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct RankFocusNeuronsOutput {
+    pub success: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub neurons: Option<Vec<RankedNeuronJson>>,
+    /// Neurons with high error but very low impact - candidates for removal
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub removal_candidates: Option<Vec<RemovalCandidateJson>>,
+    /// Issue #306: Coordinated structural candidates for removing constant-value neurons.
+    /// When a hidden neuron has near-zero activation variance (constant output), it can be
+    /// removed and its effect folded into bias adjustments for downstream neurons.
+    /// Each candidate contains:
+    /// - A RemoveNeuron operation for the constant neuron
+    /// - SetBias operations for all downstream neurons with adjusted biases
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub constant_neuron_removals: Option<Vec<CoordinatedStructuralCandidateJson>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub max_output_error: Option<f32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub processed_neurons: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub total_neurons: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub duration_ms: Option<u64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error: Option<String>,
+    /// Structured error classification for retry decisions (Issue #651).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub error_kind: Option<DiscoveryErrorKind>,
+    /// Whether this error is typically worth retrying (Issue #651).
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub retryable: Option<bool>,
+}
+
+// ============================================================================
+// Conversion helpers
+// ============================================================================
+
+pub(crate) fn synapse_diagnostics_json(
+    summaries: &[analysis_mod::shared::SynapseNoCandidateSummary],
+) -> Option<Vec<SynapseDiagnosticJson>> {
+    if summaries.is_empty() {
+        return None;
+    }
+    Some(
+        summaries
+            .iter()
+            .map(|summary| SynapseDiagnosticJson {
+                target_neuron_uuid: summary.target_uuid.clone(),
+                reason: match summary.reason {
+                    analysis_mod::shared::SynapseNoCandidateReason::NoEligibleSources => {
+                        SynapseDiagnosticReasonJson::NoEligibleSources
+                    }
+                    analysis_mod::shared::SynapseNoCandidateReason::NoDiagnostics => {
+                        SynapseDiagnosticReasonJson::NoDiagnostics
+                    }
+                    analysis_mod::shared::SynapseNoCandidateReason::NoSamples => {
+                        SynapseDiagnosticReasonJson::NoSamples
+                    }
+                    analysis_mod::shared::SynapseNoCandidateReason::ZeroImprovement => {
+                        SynapseDiagnosticReasonJson::ZeroImprovement
+                    }
+                    analysis_mod::shared::SynapseNoCandidateReason::BelowThreshold => {
+                        SynapseDiagnosticReasonJson::BelowThreshold
+                    }
+                },
+                evaluated_candidates: summary.evaluated_candidates,
+                candidates_with_samples: summary.candidates_with_samples,
+                target_record_count: summary.target_record_count,
+                detail: summary
+                    .detail
+                    .as_ref()
+                    .map(|detail| SynapseDiagnosticDetailJson {
+                        source_neuron_uuid: detail.source_uuid.clone(),
+                        sample_count: detail.sample_count,
+                        source_record_count: detail.source_record_count,
+                        improved_count: detail.improved_count,
+                        worsened_count: detail.worsened_count,
+                        expected_creature_error_reduction: detail.expected_improvement,
+                        threshold: detail.threshold,
+                        suggested_weight: detail.suggested_weight,
+                    }),
+            })
+            .collect(),
+    )
+}
+
+pub(crate) fn neuron_diagnostics_json(
+    summaries: &[analysis_mod::shared::NeuronNoCandidateSummary],
+) -> Option<Vec<NeuronDiagnosticJson>> {
+    if summaries.is_empty() {
+        return None;
+    }
+    Some(
+        summaries
+            .iter()
+            .map(|summary| NeuronDiagnosticJson {
+                target_neuron_uuid: summary.target_uuid.clone(),
+                reason: match summary.reason {
+                    analysis_mod::shared::NeuronNoCandidateReason::NoEligibleSources => {
+                        NeuronDiagnosticReasonJson::NoEligibleSources
+                    }
+                    analysis_mod::shared::NeuronNoCandidateReason::NoDiagnostics => {
+                        NeuronDiagnosticReasonJson::NoDiagnostics
+                    }
+                    analysis_mod::shared::NeuronNoCandidateReason::NoSamples => {
+                        NeuronDiagnosticReasonJson::NoSamples
+                    }
+                    analysis_mod::shared::NeuronNoCandidateReason::NotEnoughActivations => {
+                        NeuronDiagnosticReasonJson::NotEnoughActivations
+                    }
+                    analysis_mod::shared::NeuronNoCandidateReason::WeightDegenerate => {
+                        NeuronDiagnosticReasonJson::WeightDegenerate
+                    }
+                    analysis_mod::shared::NeuronNoCandidateReason::BelowThreshold => {
+                        NeuronDiagnosticReasonJson::BelowThreshold
+                    }
+                    analysis_mod::shared::NeuronNoCandidateReason::HiddenNeuronFiltered => {
+                        NeuronDiagnosticReasonJson::HiddenNeuronFiltered
+                    }
+                    analysis_mod::shared::NeuronNoCandidateReason::InputNeuronFiltered => {
+                        NeuronDiagnosticReasonJson::InputNeuronFiltered
+                    }
+                    analysis_mod::shared::NeuronNoCandidateReason::ConstantNeuronFiltered => {
+                        NeuronDiagnosticReasonJson::ConstantNeuronFiltered
+                    }
+                },
+                evaluated_sources: summary.evaluated_sources,
+                sources_with_samples: summary.sources_with_samples,
+                target_record_count: summary.target_record_count,
+                detail: summary
+                    .detail
+                    .as_ref()
+                    .map(|detail| NeuronDiagnosticDetailJson {
+                        source_neuron_uuid: detail.source_uuid.clone(),
+                        orientation: detail.orientation.clone(),
+                        sample_count: detail.sample_count,
+                        improved_count: detail.improved_count,
+                        worsened_count: detail.worsened_count,
+                        expected_creature_error_reduction: detail.expected_improvement,
+                        threshold: detail.threshold,
+                        outgoing_weight: detail.outgoing_weight,
+                    }),
+            })
+            .collect(),
+    )
+}
