+++ tests/infrastructure/main.rs
@@ -19,4 +19,5 @@ mod issue_776_hashset_hashmap_iteration;
 mod issue_832_deadlock_stress;
 mod issue_836_deadlock_abort;
 mod issue_837_lock_contention_tracing;
+mod issue_874_module_split_backward_compat;
 mod observability;
